Post-process the grade module metadata to use the shadow jar deps
By default, the gradle module metadata points at the unshaded jar and
dependencies. Modify the resulting JSON to instead use the dependencies
from the shadow jar for the `runtimeElements` variant

Note that while gradle has an API for modifying variants
(i.e. `AdhocComponentWithVariants`), it does not provide an ergonomic
means for modifying the dependencies, so use post-processing similar
to what we already do for the POM file